edit button status borrow and path

+++ src/app/login/login.component.ts
@@ -1,16 +1,16 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthService, GoogleLoginProvider } from 'angular-6-social-login';
-import { forkJoin } from 'rxjs';
-import { MacinoddsApiService } from 'src/app/service/macinodds-api.service';
-import * as JWT from 'jwt-decode';
-import { User } from '../shared/user';
-
+import { Component, OnInit, Input } from "@angular/core";
+import { Router } from "@angular/router";
+import { AuthService, GoogleLoginProvider } from "angular-6-social-login";
+import { forkJoin } from "rxjs";
+import { MacinoddsApiService } from "src/app/service/macinodds-api.service";
+import * as JWT from "jwt-decode";
+import { User } from "../shared/user";
+import { CheckRoleTokenService } from "../service/check-role-token.service";
 
 @Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrls: ['./login.component.css']
+  selector: "app-login",
+  templateUrl: "./login.component.html",
+  styleUrls: ["./login.component.css"]
 })
 export class LoginComponent implements OnInit {
   //test
@@ -21,56 +21,62 @@ export class LoginComponent implements OnInit {
     private socialAuthService: AuthService,
     private macinoddsService: MacinoddsApiService,
     private route: Router,
-  ) { }
+    private checkRoleToken: CheckRoleTokenService
+  ) {}
 
   ngOnInit() {
-    sessionStorage.clear();
-    localStorage.clear();
-    localStorage.removeItem('userResult');
-
+    if (this.macinoddsService.isAuthenticated()) {
+      if (this.checkRoleToken.checkRoleByToken() === "admin") {
+        this.route.navigate(["/admin"]);
+      } else {
+        this.route.navigate(["/user"]);
+      }
+    } else {
+      sessionStorage.clear();
+      localStorage.clear();
+      localStorage.removeItem("userResult");
+    }
+    
   }
-
 
   oddsSignIn() {
     let socialPlatformProvider;
     socialPlatformProvider = GoogleLoginProvider.PROVIDER_ID;
-    this.socialAuthService.signIn(socialPlatformProvider).then(
-      (userData) => {
-        if (this.isOddsTeam(userData.email)) {
-          this.loginGoogle(userData.idToken)
-          localStorage.setItem('Username', userData.name);
-          localStorage.setItem('email', userData.email);
-          localStorage.setItem('image', userData.image);
-        }
+    this.socialAuthService.signIn(socialPlatformProvider).then(userData => {
+      if (this.isOddsTeam(userData.email)) {
+        this.loginGoogle(userData.idToken);
+        localStorage.setItem("Username", userData.name);
+        localStorage.setItem("email", userData.email);
+        localStorage.setItem("image", userData.image);
       }
-    );
+    });
   }
 
   loginGoogle(idToken: string) {
     localStorage.clear();
     this.macinoddsService.getLoginGoogle(idToken).subscribe(res => {
-      sessionStorage.setItem('token', 'Bearer ' + res.token);
+      sessionStorage.setItem("token", "Bearer " + res.token);
       let decode = JWT(res.token);
       const decodeToString = JSON.stringify(decode);
       const decodeNew = JSON.parse(decodeToString);
       const role = decodeNew.role;
-      localStorage.setItem('userId', decodeNew.id);
-      localStorage.setItem('role', role);
+      localStorage.setItem("userId", decodeNew.id);
+      localStorage.setItem("role", role);
       if (res.firstLogin) {
-        this.route.navigate(['/register'])
+        this.route.navigate(["/register"]);
       } else {
-        if (localStorage.getItem('role') == 'admin') {
-          this.route.navigate(['/admin']);
+        if (localStorage.getItem("role") == "admin") {
+          this.route.navigate(["/admin"]);
         } else {
-          this.route.navigate(['/user']);
+          this.route.navigate(["/user"]);
         }
       }
-    })
+    });
   }
 
   private isOddsTeam(email: string): boolean {
     const host = email.slice(-10);
-    if (host !== '@odds.team') {
+    if (host !== "@odds.team") {
       alert(`Sorry, account isn't Odds Team.`);
       return false;
     }
